os-xenapi: Failed to call Dom0 plugin glance.py
There are many errors in the patchs which passed our XenServer CI,
we can see there are some errors in nova.
It is caused by the exception which we throw out in the os-xenapi
glance.py. Nova compute service couldn't recognize it.
Add a exception which computer service can process

Change-Id: Ibe30a76f278264e0ddd1fc29d533fdeaab0a13a5