XenAPI-Kolla: Gather domU's vifs in facts
This commit is to gather domU's vifs in facts. So that the compute
VM's interfaces' vif information will be included in the xenapi_facts.
When deploy OpenStack on XenServers, the vif information can be used
by deployment scripts (e.g. in Kolla-ansible, it can use the vif's
bridge to support provider networks by setting the bridge mappings).

Change-Id: I9a6bebe19ed488bb2173d5dc2daa14e236411243