Use module name (__name__) for getLogger()
We usually use a period-separated hierarchical module name as
the logger's name. But we have three scripts are using a hardcoded
name. This commit is to change them to use __name__ to make them
consistent with others. By using the period-separated hierarchical
module name, it's easier to know which module generated the log
lines so it's also easier to get the file name.

Change-Id: I373ce9b44caf2b9baffb753e6db7ad8789219f4b